+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ## 2.19 branch
 ### 2.19.2
 * added `MockTrait::getOriginClassNameOrFail()` method;
+* fixed little bug for `MockTrait::getOriginClassName()` method;
 * changed `npm-asset/fortawesome--fontawesome-free` package with
     `fortawesome/font-awesome`.